+++ panel/widgets/misc.py
@@ -16,10 +16,10 @@ from ..depends import depends
 from ..io.notebook import push
 from ..io.state import state
 from ..models import (
-    Audio as _BkAudio, VideoStream as _BkVideoStream, Progress as _BkProgress,
-    FileDownload as _BkFileDownload
+    Audio as _BkAudio, VideoStream as _BkVideoStream, FileDownload as _BkFileDownload
 )
 from .base import Widget
+from .indicators import Progress # noqa
 
 
 class _MediaBase(Widget):
@@ -130,32 +130,6 @@ class VideoStream(Widget):
             (self, root, doc, comm) = state._views[ref]
             if comm and 'embedded' not in root.tags:
                 push(doc, comm)
-
-
-class Progress(Widget):
-
-    active = param.Boolean(default=True, doc="""
-        If no value is set the active property toggles animation of the
-        progress bar on and off.""")
-
-    bar_color = param.ObjectSelector(default='success', objects=[
-        'primary', 'secondary', 'success', 'info', 'danger', 'warning',
-        'light', 'dark'])
-
-    max = param.Integer(default=100, doc="The maximum value of the progress bar.")
-
-    value = param.Integer(default=None, bounds=(0, 100), doc="""
-        The current value of the progress bar. If set to None the progress
-        bar will be indeterminate and animate depending on the active
-        parameter.""")
-
-    _rename = {'name': None}
-
-    _widget_type = _BkProgress
-
-    @depends('max', watch=True)
-    def _update_value_bounds(self):
-        self.param.value.bounds = (0, self.max)
 
 
 class FileDownload(Widget):
